+++ javacutil/src/main/java/org/checkerframework/javacutil/TreeUtils.java
@@ -1684,6 +1684,27 @@ public final class TreeUtils {
   }
 
   /**
+   * Returns the body of the case statement if it is of the form {@code case <expression> ->
+   * <expression>}. This method should only be called if {@link CaseTree#getStatements()} returns
+   * null.
+   *
+   * @param caseTree the case expression to get the body from
+   * @return the body of the case tree
+   */
+  public static @Nullable Tree caseTreeGetBody(CaseTree caseTree) {
+    try {
+      Method method = CaseTree.class.getDeclaredMethod("getBody");
+      return (Tree) method.invoke(caseTree);
+    } catch (NoSuchMethodException
+        | IllegalAccessException
+        | IllegalArgumentException
+        | InvocationTargetException e) {
+      // Just assume that the case tree is of the form case <expression> : statement(s)
+      return null;
+    }
+  }
+
+  /**
    * Returns true if the given method/constructor invocation is a varargs invocation.
    *
    * @param tree a method/constructor invocation